refactor: move subprocess/socket tests to tests/integration/
Unit tests must have no subprocess, no real sockets, no Docker
(per testing-taxonomy rule). Tests using git subprocess or real
socket binding are integration tests — they were in tests/unit/
and only failed in CI because the environment differs (no git
identity, Linux SO_REUSEADDR semantics).

Moved 31 tests to tests/integration/, 268 pure tests remain in
tests/unit/. Added test-integration target to Makefile/Justfile.
make check runs unit only (fast, CI-safe). Integration tests run
separately with make test-integration.
