+++ Makefile
@@ -1,6 +1,6 @@
 # Quality infrastructure — generated by kinfra init
 
-.PHONY: lint quality test-unit check setup test-e2e fix
+.PHONY: lint quality test-unit test-integration check setup test-e2e fix
 
 lint:
 	uv run ruff check src/ tests/
@@ -10,6 +10,9 @@ quality:
 
 test-unit:
 	uv run pytest tests/unit
+
+test-integration:
+	uv run pytest tests/integration
 
 test-e2e:
 	uv run pytest tests/e2e
