+++ tests/integration/test_cli_done.py
